[ALPS06451101] Remove redundant casting causing errors
Maximum BA size was 256 before, but in EHT it will be increased to 1024.
8-bit integer cannot hold the value, casting the value to uint8 will
cause problem.

MTK-Commit-Id: b7ff508a98ba84d00b68d0622aedf487052d5f6f

CR-Id: ALPS06451101
Feature: [Module]Wi-Fi Driver

+++ common/wlan_lib.c
@@ -6828,11 +6828,9 @@ void wlanInitFeatureOption(IN struct ADAPTER *prAdapter)
 	prWifiVar->ucP2pGcEht = (uint8_t)
 		wlanCfgGetUint32(prAdapter, "P2pGcEHT",
 				FEATURE_ENABLED);
-	prWifiVar->u2RxEhtBaSize = (uint8_t)
-		wlanCfgGetUint32(prAdapter, "RxEhtBaSize",
+	prWifiVar->u2RxEhtBaSize = wlanCfgGetUint32(prAdapter, "RxEhtBaSize",
 			WLAN_EHT_MAX_BA_SIZE);
-	prWifiVar->u2TxEhtBaSize = (uint8_t)
-		wlanCfgGetUint32(prAdapter, "TxEhtBaSize",
+	prWifiVar->u2TxEhtBaSize = wlanCfgGetUint32(prAdapter, "TxEhtBaSize",
 			WLAN_EHT_MAX_BA_SIZE);
 	prWifiVar->ucEhtAmsduInAmpduRx = (uint8_t) wlanCfgGetUint32(prAdapter,
 		"EhtAmsduInAmpduRx", FEATURE_ENABLED);
